+++ src/Numeric/Units/Dimensional/Dimensions/TermLevel.hs
@@ -1,5 +1,10 @@
 {-# OPTIONS_HADDOCK not-home, show-extensions #-}
 
+{-# LANGUAGE BangPatterns #-}
+{-# LANGUAGE DefaultSignatures #-}
+{-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE DeriveGeneric #-}
+
 {- |
    Copyright  : Copyright (C) 2006-2015 Bjorn Buckwalter
    License    : BSD3
@@ -17,7 +22,7 @@ module Numeric.Units.Dimensional.Dimensions.TermLevel
   -- * Type
   Dimension'(..),
   -- * Access to Dimension of Dimensional Values
-  HasDimension(..),
+  HasDimension(..), HasDynamicDimension(..),
   -- * Dimension Arithmetic
   (*), (/), (^), recip,
   -- * Synonyms for Base Dimensions
@@ -28,25 +33,44 @@ module Numeric.Units.Dimensional.Dimensions.TermLevel
 )
 where
 
+import Control.DeepSeq
+import Data.Data
 import Data.Monoid (Monoid(..))
-import Prelude (id, (+), (-), Int, Show, Eq, Ord)
+import GHC.Generics
+import Prelude (id, (+), (-), (.), Int, Show, Eq, Ord, Maybe(..))
 import qualified Prelude as P
 
 -- | A physical dimension, encoded as 7 integers, representing a factorization of the dimension into the
 -- 7 SI base dimensions. By convention they are stored in the same order as 
 -- in the 'Numeric.Units.Dimensional.Dimensions.TypeLevel.Dimension' data kind.
 data Dimension' = Dim' !Int !Int !Int !Int !Int !Int !Int 
-  deriving (Show, Eq, Ord)
+  deriving (Show, Eq, Ord, Data, Generic, Typeable)
+
+instance NFData Dimension' where
+  rnf !_ = () -- The Dimension' constructor is already fully strict.
 
 -- | The monoid of dimensions under multiplication.
 instance Monoid Dimension' where
   mempty = dOne
   mappend = (*)
 
+-- | Dimensional values, or those that are only possibly dimensional, inhabit this class,
+-- which allows access to a term-level representation of their dimension.
+class HasDynamicDimension a where
+  -- | Gets the 'Dimension'' of a dynamic dimensional value, or 'Nothing' if it does not represent
+  -- a dimensional value of any 'Dimension'.
+  --
+  -- A default implementation is available for types that are also in the `HasDimension` typeclass.
+  dynamicDimension :: a -> Maybe Dimension'
+  default dynamicDimension :: (HasDimension a) => a -> Maybe Dimension'
+  dynamicDimension = Just . dimension
+
 -- | Dimensional values inhabit this class, which allows access to a term-level representation of their dimension.
-class HasDimension a where 
+class HasDynamicDimension a => HasDimension a where 
   -- | Obtains a term-level representation of a value's dimension.
   dimension :: a -> Dimension'
+
+instance HasDynamicDimension Dimension' where
 
 instance HasDimension Dimension' where
   dimension = id