+++ src/Numeric/Units/Dimensional/Variants.hs
@@ -27,6 +27,7 @@ module Numeric.Units.Dimensional.Variants
 )
 where
 
+import Control.DeepSeq
 import Data.Data
 import GHC.Generics
 
@@ -35,6 +36,8 @@ import GHC.Generics
 data Metricality = Metric    -- ^ Capable of receiving a metric prefix.
                  | NonMetric -- ^ Incapable of receiving a metric prefix.
   deriving (Eq, Ord, Data, Typeable, Generic)
+
+instance NFData Metricality where -- instance is derived from Generic instance
 
 {-
 The variety 'v' of 'Dimensional'
@@ -47,6 +50,8 @@ and quantities. It must be one of the following:
 data Variant = DQuantity         -- ^ The value is a quantity.
              | DUnit Metricality -- ^ The value is a unit, possibly a 'Metric' unit.
   deriving (Eq, Ord, Data, Typeable, Generic)
+
+instance NFData Variant where -- instance is derived from Generic instance
 
 {-
 We will reuse the operators and function names from the Prelude.